chore(planner): Convert AST into logical plan

+++ pkg/engine/planner/physical/limit.go
@@ -10,9 +10,9 @@ type Limit struct {
 	id string
 
 	// Offset specifies how many initial rows should be skipped.
-	Offset uint64
+	Offset uint32
 	// Limit specifies how many rows should be returned in total.
-	Limit uint64
+	Limit uint32
 }
 
 // ID implements the [Node] interface.
